Refactor/reorganize lexer functions, add free functions and pass the norminette
- Refactor/reorganize lexer functions
	Create two files: ft_token_list.c and ft_get_token_len.c
- Add "ft_free_token_list" function to free the token list
- Pass the NORMINETTE on files in srcs/ and incs/
