+++ incs/lexer.h
@@ -6,14 +6,14 @@
 /*   By: jahuang <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/10 21:03:52 by jahuang           #+#    #+#             */
-/*   Updated: 2022/01/10 21:56:08 by jahuang          ###   ########.fr       */
+/*   Updated: 2022/01/13 18:05:04 by jahuang          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LEXER_H
 # define LEXER_H
 
-enum	char_type {
+enum	e_char_type {
 	CHAR_GENERAL = -1,
 	CHAR_PIPE = '|',
 	CHAR_GREATER = '>',
@@ -23,7 +23,7 @@ enum	char_type {
 	CHAR_D_QUOTE = '"',
 };
 
-enum	token_type {
+enum	e_token_type {
 	TOKEN_WORD = 0,
 	TOKEN_PIPE,
 	TOKEN_GREATER,
@@ -32,13 +32,16 @@ enum	token_type {
 	TOKEN_LESSLESSER,
 };
 
-typedef struct	s_token {
-	char *value;
-	int	token_type;
-	struct s_token *prev;
-	struct s_token *next;
+typedef struct s_token {
+	char			*value;
+	int				token_type;
+	struct s_token	*prev;
+	struct s_token	*next;
 }		t_token;
 
 t_token		*ft_lexer(char *input);
+t_token		*ft_fill_list(t_token *token_list, char *input, int token_len);
+void		ft_free_token_list(t_token *token_list);
+int			ft_get_token_len(char *input);
 
 #endif
