+++ srcs/minishell.c
@@ -6,19 +6,19 @@
 /*   By: antton-t <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/13 19:32:58 by antton-t          #+#    #+#             */
-/*   Updated: 2022/01/10 22:47:26 by jahuang          ###   ########.fr       */
+/*   Updated: 2022/01/13 18:01:36 by jahuang          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
 /* function to print out token_list for testing purpous. */
-static void ft_print_token_list(t_token *token_list)
+static void	ft_print_token_list(t_token *token_list)
 {
-	t_token *holder;
+	t_token	*holder;
 
 	holder = token_list;
-	while(holder)
+	while (holder)
 	{
 		printf("token type: %d\n", holder->token_type);
 		printf("token value: %s\n\n", holder->value);
@@ -27,11 +27,11 @@ static void ft_print_token_list(t_token *token_list)
 	return ;
 }
 
-static void ft_print_title(void)
+static void	ft_print_title(void)
 {
-	char buffer[2048 + 1];
-	int	fd;
-	int	ret;
+	char	buffer[2048 + 1];
+	int		fd;
+	int		ret;
 
 	fd = open("./others/prompt_string.txt", O_RDONLY);
 	ret = read(fd, buffer, 2048);
@@ -39,22 +39,29 @@ static void ft_print_title(void)
 	printf("%s\n", buffer);
 }
 
-int main(int argc, char **argv, char **env)
+int	main(int argc, char **argv, char **env)
 {
-	char *input;
-	char *prompt;
-	(void)	argc;
-	(void)	argv;
-	(void)	env;
+	char	*input;
+	char	*prompt;
 	t_token	*token_list;
 
+	(void)argc;
+	(void)argv;
+	(void)env;
 	prompt = "\n|( o)═( o)| >";
 	ft_print_title();
 	while (1)
 	{
 		input = readline(prompt);
+		if (!input)
+		{
+			printf("Banana~~~!\n");
+			exit(0);
+		}
 		printf("User input: %s\n", input);
 		token_list = ft_lexer(input);
 		ft_print_token_list(token_list);
+		ft_free_token_list(token_list);
+		free(input);
 	}
 }
